refactor(#3272): extract struct-field-exports from index.ts (refs #3182)
__get_field_*/__set_field_*/__struct_field_names exports + getter dispatch
builders (buildNestedIfElse/buildGetterExtract) moved verbatim; index.ts
imports back the 3 driver entry points. Byte-identity IDENTICAL (39/39), tsc 0.

✓
